rename CAM2CameraDatabaseAPIClient
