+++ CAM2CameraDatabaseAPIClient/camera.py
