+++ CAM2CameraDatabaseAPIClient/client.py
