+++ CAM2CameraDatabaseAPIClient/config.py
