+++ CAM2CameraDatabaseAPIClient/tests/test_camera.py
